@@ -1,20 +1,27 @@
 {
-  "name": "svelte-zoom",
-  "version": "1.0.16",
+  "name": "@pfzone/svelte-zoom",
+  "version": "1.0.0",
+  "type": "module",
   "description": "Nearly native image scaling in web (mobile/desktop), rich in functionality. This will make your mobile users happy",
   "svelte": "src/index.svelte",
   "main": "dist/index.js",
-  "module": "dist/index.mjs",
+  "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "svelte": "./src/index.svelte",
+      "default": "./dist/index.js"
+    },
+    "./index.svelte": "./src/index.svelte"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "rollup -c",
-    "pretty": "prettier --config .prettierrc --write --svelte-sort-order styles-markup-scripts \"rollup.config.js\" \"src/*.svelte\" \"src/*.js\" \"package.json\" \"*.md\" ",
+    "build": "vite build",
+    "pretty": "prettier --config .prettierrc --write \"vite.config.ts\" \"src/*.svelte\" \"src/*.ts\" \"package.json\" \"*.md\"",
     "go": "npm-run-all pretty build",
     "prepublishOnly": "npm run build"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/vaheqelyan/svelte-zoom.git"
+    "url": "git+https://github.com/pfzone/svelte-zoom.git"
   },
   "author": "Vahe Araqelyan",
   "license": "MIT",
@@ -34,14 +41,18 @@
     "mouse"
   ],
   "homepage": "https://github.com/vaheqelyan/svelte-zoom#readme",
+  "peerDependencies": {
+    "svelte": "^5.20.0"
+  },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tsconfig/svelte": "^5.0.4",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.1.1",
-    "prettier-plugin-svelte": "^1.2.0",
-    "rollup": "^2.26.9",
-    "rollup-plugin-svelte": "^6.0.0",
-    "svelte": "^3.24.1"
+    "prettier": "^3.5.2",
+    "prettier-plugin-svelte": "^3.3.3",
+    "svelte": "^5.20.5",
+    "typescript": "^5.7.3",
+    "vite": "^6.2.0"
   },
   "files": [
     "src",
